Move MultiLineJaggedTextTests into navigation package
