+++ richtextfx/src/integrationTest/java/org/fxmisc/richtext/keyboard/NavigationTests.java
@@ -43,53 +43,6 @@ public class NavigationTests {
         }
     }
 
-    public class MultiLineJaggedTextTests extends InlineCssTextAreaAppTest {
-
-        String threeLinesOfText = "Some long amount of text to take up a lot of space in the given area.";
-
-        @Override
-        public void start(Stage stage) throws Exception {
-            super.start(stage);
-            stage.setWidth(200);
-            area.replaceText(threeLinesOfText);
-            area.setWrapText(true);
-        }
-
-        private void waitForMultiLineRegistration() throws TimeoutException {
-            // When the stage's width changes, TextFlow does not properly handle API calls to a
-            //  multi-line paragraph immediately. So, wait until it correctly responds
-            //  to the stage width change
-            Future<Void> textFlowIsReady = WaitForAsyncUtils.asyncFx(() -> {
-                while (area.getParagraphLinesCount(0) != 3) {
-                    sleep(10);
-                }
-            });
-            WaitForAsyncUtils.waitFor(5, TimeUnit.SECONDS, textFlowIsReady);
-        }
-
-        @Test
-        public void pressingDownMovesCaretToNextLine() throws TimeoutException {
-            waitForMultiLineRegistration();
-
-            area.moveTo(27);
-
-            push(DOWN);
-
-            assertEquals(57, area.getCaretPosition());
-        }
-
-        @Test
-        public void pressingUpMovesCaretToPrevLine() throws TimeoutException {
-            waitForMultiLineRegistration();
-
-            area.moveTo(66);
-
-            push(UP);
-
-            assertEquals(36, area.getCaretPosition());
-        }
-    }
-
     public class ViewportTests extends InlineCssTextAreaAppTest {
 
         @Override
